requirements.txt
geemap
imageio
Pillow
